Issue 159 ctidy (#160)
* ctidy - first fix

* fixes from ctidy

* ctidy last step for today

* cformat

* ctidy next part of changes

* fix

* tidy fixes part 3

* ctidy test update

* get rid of magic numbers (except tests)

* Cformat

* fix: remove redundant double std::move in field.cpp

* Replace sprintf with snprintf and handle truncation in qry.cpp

* fix: change static_cast<int> to static_cast<size_t> for revRead call

---------

+++ src/rdb/cmdFieldAccess.cpp
@@ -56,7 +56,7 @@ bool GetPosCmd::execute(CommandContext &ctx) {
   else if (value.type() == typeid(int))
     std::cout << std::any_cast<int>(value) << "\n";
   else if (value.type() == typeid(uint8_t))
-    std::cout << std::any_cast<uint8_t>(value) << "\n";
+    std::cout << static_cast<unsigned int>(std::any_cast<uint8_t>(value)) << "\n";
   else if (value.type() == typeid(float))
     std::cout << std::any_cast<float>(value) << "\n";
   else if (value.type() == typeid(double))
